packages/lemieux-glossary/pyproject.toml
[project]
name = "lemieux-glossary"
version = "0.1.0"
description = "Bilingual (EN/FR) glossary of hockey analytics terms with formulas, caveats, and sources"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=8.3"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true

[tool.setuptools.package-data]
"lemieux.glossary" = ["*.yaml"]
